requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.12.1
async-timeout==4.0.3
attrs==25.4.0
blinker==1.9.0
cachetools==5.5.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.1.8
crewai==0.5.0
cryptography==46.0.5
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.3.1
faiss-cpu==1.9.0.post1
frozenlist==1.8.0
gitdb==4.0.12
gitpython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.1.0
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.2.post1
langchain-text-splitters==0.3.4
langsmith==0.0.87
markdown-it-py==3.0.0
markupsafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
multidict==6.7.1
mypy-extensions==1.1.0
narwhals==2.17.0
numpy==1.26.4
openai==1.58.1
orjson==3.11.5
packaging==23.2
pandas==2.3.3
pdfminer-six==20231228
pdfplumber==0.11.4
pillow==11.3.0
pip==26.0.1
propcache==0.4.1
protobuf==5.29.6
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.11.0
pydeck==0.9.1
pygments==2.19.2
pypdfium2==5.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
pyyaml==6.0.3
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.27.1
setuptools==44.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.48
streamlit==1.40.2
tenacity==8.5.0
tiktoken==0.5.2
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
yarl==1.22.0
